AddAlarm: Follow the 2nd variation in the paper design
The design describes the app both in a list and in a flickable page.
I chose the list version since it gives more space for the list
and I think it made it more compact and comprehensive to look at.

I do find it disturbing to have a map take ~50% of space while being persistent
even when you're going to use it once.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>