Babana: Implement Notify Page
It links to an alarm "item" and would be responsible for creating a structured
list of arbitrarily chosen recipients.

This would feature a simple text editor with a 160-character limit, another page
containing a full exhaustive list of the phonebook entires with mobile numbers.

The SMS sending is not part of this component.

lol.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
(cherry picked from commit 49a1b51b25f42fae6df8f84ace0c34dc79ff437a)
Though this is a cherry-picked commit on a branch which may have been
already removed, this is not of the same version.

It has been ported to NativeBase and uses this branch's structure and such.
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>