Babana: Implement Notify Page
It links to an alarm "item" and would be responsible for creating a structured
list of arbitrarily chosen recipients.

This would feature a simple text editor with a 160-character limit, another page
containing a full exhaustive list of the phonebook entires with mobile numbers.

The SMS sending is not part of this component.

lol.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>