Generalize RingtoneSettings key; Acknowledge the is_null argument
Restructure the code into two more functions to make the
componentWillMount() less messy.
We also remove the excess console.log() and remove the .catch()
from loadDefaultRingtone() since it's not a promise.

Finally, add the is_null argument is added and stores the
stringified version of the state instead of the URI component only.
This allows us to store null values for the URI which is not possible as a string.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>