babana-ringtone: Reset MediaPlayer after stopping the playback
As it turns out, you may not call start() after a stop().
Otherwise, the next start() call will not initiate playback.

There are at least 3 options:
1. Instead of stopping, just pause.
2. Reset and Reload the stream (ringtone in out case)
3. Read the MediaPlayer state diagram and call prepare() before start().

The intelligible option should be 3, but it causes glitches like still playing after it stopped
and the playback from the next call of start() causes the playback to start
and stutter from some point in the tune until it ends or at the next loop.

The first option is nice, but the problematic function is called "stopRingtone" for a reason.

So the last option, the second option, is used. It works just as expected.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>