Pass through SMS Sending to default messaging apps
Google wants to restrict the SEND_SMS permission for security.
Due to recent changes on the app permission policy, there should be at least 3
ways to allow the app to stay in the play store: Remove the SMS Functionality,
Petition for exclusion, or use an alternative way to send SMS without using the
SEND_SMS permission.

To be honest, I don't like to remove the SMS Functionality. It's the only thing
I've contributed to the group :( lol.
So to the alternatives, it's possible to temporarily set the default messaging
app to Babana then send the messages or to pass the recipient list and the
message body to the default messaging app then the user must manually send the
sms.

Either way needs manual interaction of the user, about 3-5 taps for the "default
messaging alternative" and about 1-2 taps for the "pass to the default messaging
app alternative".

And so, I've decided to just pass it through the default messaging app. It's
simple enough to write.

This patch does that, basing the code upon the react-native-communications
npm module.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>